tracking-dashboard/backend/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
crontab==1.0.1
Flask==2.3.2
freezegun==1.2.2
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
psycogreen==1.0.2
psycopg2==2.9.6
python-box==7.0.1
python-dateutil==2.8.2
PyYAML==6.0
redis==4.5.5
requests==2.31.0
rq==1.15.0
rq-scheduler==0.13.1
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.5
zope.event==4.6
zope.interface==6.0
